Allow command forms for Java options
In order to enhance the classpath + Java options when running TLC,
this patch adds a way to mix static content with dynamic ones when
setting `CFG_JAVA_OPTIONS`.

The dynamism is inspired by the `$(...)` form from Bash, anything
inside these parentheses will be run as a `spawn` command.

The more general use case is to be able to add anything to the TLC
options. My particular use case is to be able to add modules on-demand
when using VSCode, e.g. when using

```json
{... ...
 ... ...
 "tlaplus.java.options": "$(tladeps --tladeps-vscode --tladeps-raw-deps '{io.github.pfeodrippe/tladeps-edn-module {:mvn/version \"0.5.0\", :tladeps/edn-namespaces [\"tla-edn-module.core\"]}}')"
 ... ...
 ... ...
}
```

the command `tladeps ...` would be run and replaced with its
result (it's a very simple regex).

Let me know what you think, hopefully this is generic enough, but if
you want something more specific for the `tladeps` case, I can do it
too o/

Signed-off-by: Paulo Feodrippe <paulo.feodrippe@Paulos-MacBook-Pro.local>