Partial reapply c58ce91.
When the user has set the config:
```
"editor.codeActionsOnSave": {
      "source": "explicit"
    }
```
this runs debugging#smokeTestSpec function which in turn will
check for the presence of the .cfg files and trigger the dialogs.

With this fix, the dialogs will show up only when the user
is trying to model check the currently selected spec.