Upgrade outdated tla2tools and CommunityModules-deps.jar.
Updating them does not affect the published/released TLA+ VS Code extension, since it is always bundled with the latest versions of tla2tools and the CommunityModules JAR. This commit ensures that our development environment uses sufficiently up-to-date versions. Updates are done infrequently to avoid unnecessary growth of the git repository (while integrating this into the build is possible, the effort is not justified).

[Dev]

Signed-off-by: Markus Alexander Kuppe <github.com@lemmster.de>